requirements.txt
django
djangorestframework
pyyaml
uritemplate
django-heroku
gunicorn
